refactor: lot 4 — extraire hooks Supabase des composants + règle dependency-cruiser
- Extraire useSignalements.ts depuis SignalementsBlock.tsx (auth.getUser + issue_reports CRUD)
- Extraire usePassHistory.ts depuis PassHistoryAccordion.tsx (rpc + pass_history CRUD)
- Hooks dans src/components/settings/hooks/ (pattern autorisé)
- Composants ne contiennent plus d'import supabaseClient direct
- Ajouter règle no-supabase-from-components dans .dependency-cruiser.cjs
  (src/components/ → supabaseClient = error, sauf hooks/ dédiés)
- Vérification : check:arch ✅ + typecheck ✅
