+++ .dependency-cruiser.cjs
@@ -83,6 +83,16 @@ module.exports = {
 
     // ── 4. Cross-feature internals (un règle par feature) ────────────────────────────────
     ...FEATURES.map(crossFeatureRule),
+
+    // ── 5. components/ : pas d'import direct de supabaseClient ─────────────────────────
+    // Les hooks dédiés dans components/settings/hooks/ sont autorisés (c'est le pattern cible).
+    {
+      name: 'no-supabase-from-components',
+      severity: 'error',
+      comment: 'src/components/ ne doit pas importer supabaseClient directement — utiliser un hook dédié dans hooks/ (AGENTS.md §3)',
+      from: { path: '^src/components/', pathNot: '^src/components/.*/hooks/' },
+      to: { path: 'supabaseClient' },
+    },
   ],
 
   options: {
